chore: move endpoints to fern
