+++ fern/api/definition/common/types.yml
@@ -49,6 +49,16 @@ types:
             - zohocrm
             - sfdc
             - pipedrive
+    StandardObject:
+        enum:
+            - company
+            - contact
+            - deal
+            - event
+            - lead
+            - note
+            - task
+            - user
     ResponseStatus:
         enum:
             - ok
