+++ fern/api/definition/properties.yml
@@ -1,0 +1,43 @@
+# yaml-language-server: $schema=https://raw.githubusercontent.com/fern-api/fern/main/fern.schema.json
+
+imports:
+    errors: ./common/errors.yml
+    types: ./common/types.yml
+
+types:
+    FieldDetailsType:
+        properties:
+            fieldName: string
+            fieldType: string
+            fieldDescription: string
+    GetObjectPropertiesResponse: unknown
+
+service:
+    base-path: /properties
+    auth: false
+    headers:
+        x-revert-api-token:
+            type: string
+            docs: Your official API key for accessing revert apis.
+        x-revert-t-id:
+            type: string
+            docs: The unique customer id used when the customer linked their account.
+        x-api-version:
+            type: optional<string>
+            docs: Optional Revert API version you're using. If missing we default to the latest version of the API.
+    audiences:
+        - external
+    endpoints:
+        getObjectProperties:
+            docs: Get object properties for connection
+            method: GET
+            path: /{objectName}
+            path-parameters:
+                objectName: types.StandardObject
+            request:
+                name: GetObjectPropertiesRequest
+            response: GetObjectPropertiesResponse
+            errors:
+                - errors.UnAuthorizedError
+                - errors.InternalServerError
+                - errors.NotFoundError
